requirements.txt
attrs==21.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
colorama==0.4.4
colored==1.4.2
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
emoji==1.2.0
idna==2.10
imageio==2.9.0
imantics==0.1.12
imgaug==0.4.0
jsonschema==3.2.0
kiwisolver==1.3.1
labelme2coco==0.1.2
lxml==4.6.3
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.54
paramiko==2.7.2
Pillow==8.2.0
pixellib==0.6.1
progress==1.6
progressbar==2.5
progressbar2==3.53.1
pycparser==2.20
pymongo==3.11.4
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pysftp==0.2.9
python-dateutil==2.8.1
python-dotenv==0.17.1
python-utils==2.5.6
PyWavelets==1.1.1
redis==3.5.3
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.3
Shapely==1.7.1
six==1.16.0
slug==2.0
termcolor==1.1.0
tifffile==2021.6.14
tqdm==4.62.2
urllib3==1.26.4
xmljson==0.2.1
xmltodict==0.12.0
